+++ src/commands/commandList.js
@@ -1,6 +1,8 @@
 const { PrivateKeyService } = require('../services/PrivateKeyService')
 const CreateWalletCommand = require('../commands/CreateWalletCommand')
 const CreateWalletCommandValidator = require('../validators/CreateWalletCommandValidator')
+const DepositEthCommand = require('./DepositEthCommand')
+const DepositEthCommandValidator = require('../validators/DepositEthCommandValidator')
 const GetAddressCommand = require('../commands/GetAddressCommand')
 const GetAddressCommandValidator = require('../validators/GetAddressCommandValidator')
 const logger = require('../logger')
@@ -13,6 +15,7 @@ const WithdrawCommand = require('../commands/WithdrawCommand')
 const WithdrawCommandValidator = require('../validators/WithdrawCommandValidator')
 
 const createWalletCommandValidator = new CreateWalletCommandValidator(logger)
+const depositEthCommandValidator = new DepositEthCommandValidator(logger)
 const getAddressCommandValidator = new GetAddressCommandValidator(logger)
 const orderSignCommandValidator = new OrderSignCommandValidator(logger)
 const withdrawCommandValidator = new WithdrawCommandValidator(logger)
@@ -29,6 +32,9 @@ const getAddressCommand = new GetAddressCommand(logger, tradingWalletService,
 const createWalletCommand = new CreateWalletCommand(logger, tradingWalletService,
   createWalletCommandValidator, privateKeyService, privateKeyValidator)
 
+const depositEthCommand = new DepositEthCommand(logger, tradingWalletService,
+  depositEthCommandValidator, privateKeyService, privateKeyValidator)
+
 const signCommand = new OrderSignCommand(logger, ordersignerHelper, orderSignCommandValidator,
   privateKeyService, privateKeyValidator)
 
@@ -37,6 +43,7 @@ const withdrawCommand = new WithdrawCommand(logger, tradingWalletService, withdr
 
 module.exports = {
   createWalletCommand,
+  depositEthCommand,
   getAddressCommand,
   signCommand,
   withdrawCommand,