+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Release date: 2019-02-REPLEACE_ME.
 
 ### Added
 
+- [CLI deposit ETH command](https://github.com/eidoo/hybrid-exchange-sdk/issues/15)
 - [CLI withdraw command](https://github.com/eidoo/hybrid-exchange-sdk/issues/13)
 - [Allowance](https://github.com/eidoo/hybrid-exchange-sdk/issues/7).
 - [Private key from mnemonic](https://github.com/eidoo/hybrid-exchange-sdk/issues/5).